pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "siteops"
dynamic = ["version"]
description = "Azure Site Ops - Multi-site orchestration for Azure IaC"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Microsoft Corporation"},
    {name = "digimaun"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]
keywords = ["azure", "iot", "infrastructure-as-code", "deployment", "orchestration"]
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
siteops = "siteops.cli:main"

[project.urls]
Homepage = "https://github.com/Azure/digital-ops-scale-kit"
Documentation = "https://github.com/Azure/digital-ops-scale-kit/tree/main/docs"
Repository = "https://github.com/Azure/digital-ops-scale-kit"
Issues = "https://github.com/Azure/digital-ops-scale-kit/issues"

[tool.setuptools.dynamic]
version = {attr = "siteops.__version__"}

[tool.setuptools.packages.find]
include = ["siteops*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: requires Azure credentials and real resources (deselect with '-m not integration')",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["siteops"]
omit = ["siteops/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
